+++ R/estimate_contrasts.R
@@ -148,7 +148,7 @@ estimate_contrasts.default <- function(model,
                                        predict = NULL,
                                        ci = 0.95,
                                        comparison = "pairwise",
-                                       estimate = "average",
+                                       estimate = getOption("modelbased_estimate", "typical"),
                                        p_adjust = "none",
                                        transform = NULL,
                                        method = "pairwise",
@@ -191,11 +191,13 @@ estimate_contrasts.default <- function(model,
   info <- attributes(estimated)
 
   # Table formatting
-  attr(out, "table_title") <- c(ifelse(
-    estimate == "specific",
-    "Model-based Contrasts Analysis",
-    "Marginal Contrasts Analysis"
+  attr(out, "table_title") <- c(switch(estimate,
+    specific = "Model-based Contrasts Analysis",
+    typical = "Marginal Contrasts Analysis",
+    average = "Averaged Contrasts Analysis",
+    population = "Counterfactual Contrasts Analysis (G-computation)"
   ), "blue")
+
   attr(out, "table_footer") <- .table_footer(
     out,
     by = info$contrast,