+++ R/estimate_means.R
@@ -22,9 +22,6 @@
 #' [this vignette](https://CRAN.R-project.org/package=emmeans/vignettes/transformations.html).
 #' Valid options for `predict` are:
 #'
-#' * `backend = "emmeans"`: `predict` can be `"response"`, `"link"`, `"mu"`,
-#'   `"unlink"`, or `"log"`. If `predict = NULL` (default), the most appropriate
-#'   transformation is selected (which usually is `"response"`).
 #' * `backend = "marginaleffects"`: `predict` can be `"response"`, `"link"` or
 #'   any valid `type` option supported by model's class `predict()` method (e.g.,
 #'   for zero-inflation models from package **glmmTMB**, you can choose
@@ -32,6 +29,9 @@
 #'   [glmmTMB::predict.glmmTMB]). By default, when `predict = NULL`, the most
 #'   appropriate transformation is selected, which usually returns predictions
 #'   or contrasts on the response-scale.
+#' * `backend = "emmeans"`: `predict` can be `"response"`, `"link"`, `"mu"`,
+#'   `"unlink"`, or `"log"`. If `predict = NULL` (default), the most appropriate
+#'   transformation is selected (which usually is `"response"`).
 #'
 #' `"link"` will leave the values on scale of the linear predictors.
 #' `"response"` (or `NULL`) will transform them on scale of the response
@@ -44,39 +44,66 @@
 #' @param estimate Character string, indicating the type of target population
 #' predictions refer to. This dictates how the predictions are "averaged" over
 #' the non-focal predictors, i.e. those variables that are not specified in
-#' `by` or `contrast`.
-#' - `"average"` (default): Takes the mean value for non-focal numeric
-#'   predictors and marginalizes over the factor levels of non-focal terms,
-#'   which computes a kind of "weighted average" for the values at which these
-#'   terms are hold constant. These predictions are a good representation of the
-#'   sample, because all possible values and levels of the non-focal predictors
-#'   are considered. It answers the question, "What is the predicted value for
-#'   an 'average' observation in *my data*?". Cum grano salis, it refers to
-#'   randomly picking a subject of your sample and the result you get on
-#'   average. This approach is the one taken by default in the `emmeans`
-#'   package.
-#' - `"population"`: Non-focal predictors are marginalized over the observations
-#'   in the sample, where the sample is replicated multiple times to produce
-#'   "counterfactuals" and then takes the average of these predicted values
-#'   (aggregated/grouped by the focal terms). It can be considered as
-#'   extrapolation to a hypothetical target population. Counterfactual
-#'   predictions are useful, insofar as the results can also be transferred to
-#'   other contexts (Dickerman and Hernan, 2020). It answers the question, "What
-#'   is the predicted response value for the 'average' observation in *the
-#'   broader target population*?". It does not only refer to the actual data in
-#'   your observed sample, but also "what would be if" we had more data, or if
-#'   we had data from a different sample.
+#' `by` or `contrast`. We can roughly distinguish between "modelbased" and
+#' "empirical" predictions.
+#' - `"typical"` (default): Predictions are made for observations that are
+#'   represented by a data grid, which is built from all combinations of the
+#'   predictor levels in `by` (the focal predictors). `"typical"` then takes the
+#'   mean value for non-focal numeric predictors and marginalizes over the
+#'   factor levels of non-focal predictors, which computes a kind of "weighted
+#'   average" for the values at which these terms are hold constant. These
+#'   predictions are useful for comparing defined "groups" and are still a good
+#'   representation of the sample, because all possible values and levels of the
+#'   non-focal predictors are considered (averaged over). It answers the
+#'   question, "What would be the average outcome for a 'typical' observation?",
+#'   where 'typical' refers to subjects represented by (i.e., that share the
+#'   characteristics from) the data grid. This approach is the one taken by
+#'   default in the `emmeans` package.
+#' - `"average"`: Predictions are made for each observation in the sample. Then,
+#'   the average of all predictions is calculated within all groups (or levels)
+#'   of the focal predictors defined in `by`. These predictions are the closest
+#'   representation of the sample, because `estimate = "average"` averages
+#'   across the full sample, where groups (in `by`) are not represented by a
+#'   balanced data grid, but rather the empirical distributions of the
+#'   characteristics of the sample. It answers the question, "What is the
+#'   predicted value for an average observation (from a certain group in `by`)
+#'   in my data?".
+#' - `"population"`: Each observation is "cloned" multiple times, where each
+#'   duplicate gets one of the levels from the focal predictors in `by`. We then
+#'   have one "original" and several copies of that original, each varying in
+#'   the levels of the focal predictors. Hence, the sample is replicated
+#'   multiple times to produce "counterfactuals" and then takes the average of
+#'   these predicted values (aggregated/grouped by the focal predictors). It can
+#'   be considered as extrapolation to a hypothetical target population.
+#'   Counterfactual predictions are useful, insofar as the results can also be
+#'   transferred to other contexts (Dickerman and Hernan, 2020). It answers the
+#'   question, "What is the predicted response value for the 'average'
+#'   observation in *the broader target population*?". It does not only refer to
+#'   the actual data in your observed sample, but also "what would be if" we had
+#'   more data, or if we had data from a different sample.
 #'
 #' In other words, the distinction between estimate types resides in whether
 #' the prediction are made for:
-#' - A specific "individual" from the sample (i.e., a specific combination of
-#'   predictor values): this is what is obtained when using [`estimate_relation()`]
-#'   and the other prediction functions.
-#' - An average individual from the sample: obtained with
-#'   `estimate_means(..., estimate = "average")`
-#' - The broader, hypothetical target population: obtained with
-#'   `estimate_means(..., estimate = "population")`
-#' @param backend Whether to use `"emmeans"` or `"marginaleffects"` as a backend.
+#' - *modelbased predictions* (focus lies on _predictors_), which are useful to
+#'   look at differences between typical groups, or for visualization
+#'   - A specific individual from the sample (i.e., a specific combination of
+#'     predictor values for focal and non-focal predictors): this is what is obtained
+#'     when using [`estimate_relation()`] and the other prediction functions.
+#'   - A typical individual from the sample: obtained with
+#'     `estimate_means(..., estimate = "typical")`
+#' - *empirical predictions* (focus lies on _predictions_ of the outcome), which
+#'   are useful if you want realistic predictions of your outcome, assuming that
+#'   the sample is representative for a special population (option `"average"`),
+#'   or useful for "what-if" scenarios, especially if you want to make unbiased
+#'   comparisons (G-computation, option `"population"`)
+#'   - The average individual from the sample: obtained with
+#'     `estimate_means(..., estimate = "average")`
+#'   - The broader, hypothetical target population: obtained with
+#'     `estimate_means(..., estimate = "population")`
+#'
+#' You can set a default option for the `estimate` argument via `options()`,
+#' e.g. `options(modelbased_estimate = "average")`
+#' @param backend Whether to use `"marginaleffects"` or `"emmeans"`as a backend.
 #' Results are usually very similar. The major difference will be found for mixed
 #' models, where `backend = "marginaleffects"` will also average across random
 #' effects levels, producing "marginal predictions" (instead of "conditional
@@ -97,12 +124,12 @@
 #' @param ... Other arguments passed, for instance, to [insight::get_datagrid()],
 #' to functions from the **emmeans** or **marginaleffects** package, or to process
 #' Bayesian models via [bayestestR::describe_posterior()]. Examples:
-#' - `insight::get_datagrid()`: Argument such as `length` or `range` can be used
-#'   to control the (number of) representative values.
+#' - `insight::get_datagrid()`: Argument such as `length`, `digits` or `range`
+#'   can be used to control the (number of) representative values.
 #' - **marginaleffects**: Internally used functions are `avg_predictions()` for
-#'   means and contrasts, and `avg_slope()` for slopes. Therefore, arguments
-#'   for instance like `vcov`, `transform`, `equivalence`, `slope` or even
-#'   `newdata` can be passed to those functions.
+#'   means and contrasts, and `avg_slope()` for slopes. Therefore, arguments for
+#'   instance like `vcov`, `equivalence`, `df`, `slope` or even `newdata` can be
+#'   passed to those functions.
 #' - **emmeans**: Internally used functions are `emmeans()` and `emtrends()`.
 #'   Additional arguments can be passed to these functions.
 #' - Bayesian models: For Bayesian models, parameters are cleaned using
@@ -114,6 +141,16 @@
 #' @inherit estimate_slopes details
 #'
 #' @return A data frame of estimated marginal means.
+#'
+#' @section Global Options to Customize Estimation of Marginal Means:
+#'
+#' - `modelbased_backend`: `options(modelbased_backend = <string>)` will set a
+#'   default value for the `backend` argument and can be used to set the package
+#'   used by default to calculate marginal means. Can be `"marginalmeans"` or
+#'   `"emmeans"`.
+#'
+#' - `modelbased_estimate`: `options(modelbased_estimate = <string>)` will
+#'   set a default value for the `estimate` argument.
 #'
 #' @references
 #' Dickerman, Barbra A., and Miguel A. Hernán. 2020. Counterfactual Prediction
@@ -180,7 +217,7 @@ estimate_means <- function(model,
                            by = "auto",
                            predict = NULL,
                            ci = 0.95,
-                           estimate = "average",
+                           estimate = getOption("modelbased_estimate", "typical"),
                            transform = NULL,
                            backend = getOption("modelbased_backend", "marginaleffects"),
                            verbose = TRUE,
@@ -188,7 +225,7 @@ estimate_means <- function(model,
   # validate input
   estimate <- insight::validate_argument(
     estimate,
-    c("average", "population", "specific", "sample")
+    c("typical", "population", "specific", "average")
   )
 
   if (backend == "emmeans") {
@@ -220,11 +257,13 @@ estimate_means <- function(model,
   info <- attributes(estimated)
 
   # Table formatting
-  attr(means, "table_title") <- c(ifelse(
-    estimate == "specific",
-    "Model-based Predictions",
-    "Estimated Marginal Means"
+  attr(means, "table_title") <- c(switch(estimate,
+    specific = "Model-based Predictions",
+    typical = "Estimated Marginal Means",
+    average = "Average Predictions",
+    population = "Average Counterfactual Predictions"
   ), "blue")
+
   attr(means, "table_footer") <- .table_footer(
     means,
     type = ifelse(estimate == "specific", "predictions", "means"),