+++ R/get_marginaltrends.R
@@ -32,11 +32,7 @@ get_marginaltrends <- function(model,
     datagrid <- datagrid_info <- NULL
   } else {
     # setup arguments
-    dg_args <- list(
-      model,
-      by = by,
-      verbose = FALSE
-    )
+    dg_args <- list(model, by = by, verbose = FALSE)
     # add user-arguments from "...", but remove those arguments that are already set
     dots[c("by", "verbose")] <- NULL
     dg_args <- insight::compact_list(c(dg_args, dots))
@@ -54,6 +50,9 @@ get_marginaltrends <- function(model,
   if (!is.null(datagrid)) {
     datagrid <- as.data.frame(datagrid)
   }
+
+  # remove user-arguments from "..." that will be used when calling marginaleffects
+  dots[c("by", "conf_level", "digits")] <- NULL
 
   # setup arguments again
   fun_args <- insight::compact_list(c(